refactor: move ticket creation to separate file
