+++ apps/bot/package.json
@@ -18,13 +18,13 @@
 		"@ticketer/env": "workspace:*",
 		"chalk": "^5.3.0",
 		"discord.js": "^14.15.3",
-		"tsx": "^4.17.0",
+		"tsx": "^4.19.0",
 		"typesafe-i18n": "^5.26.2",
 		"zod": "^3.23.8"
 	},
 	"devDependencies": {
 		"@eslint/eslintrc": "^3.1.0",
 		"@ticketer/eslint-config": "workspace:*",
-		"@types/node": "^22.4.1"
+		"@types/node": "^22.5.1"
 	}
 }
