+++ apps/website/package.json
@@ -20,8 +20,8 @@
 		"@vercel/speed-insights": "^1.0.12",
 		"class-variance-authority": "^0.7.0",
 		"clsx": "^2.1.1",
-		"lucide-react": "^0.428.0",
-		"next": "14.2.5",
+		"lucide-react": "^0.437.0",
+		"next": "14.2.7",
 		"next-themes": "^0.3.0",
 		"react": "^18.3.1",
 		"react-dom": "^18.3.1",
@@ -31,10 +31,10 @@
 	"devDependencies": {
 		"@eslint/compat": "^1.1.1",
 		"@eslint/eslintrc": "^3.1.0",
-		"@next/eslint-plugin-next": "^14.2.5",
+		"@next/eslint-plugin-next": "^14.2.7",
 		"@ticketer/eslint-config": "workspace:*",
-		"@types/node": "^22.4.1",
-		"@types/react": "^18.3.3",
+		"@types/node": "^22.5.1",
+		"@types/react": "^18.3.5",
 		"@types/react-dom": "^18.3.0",
 		"autoprefixer": "^10.4.20",
 		"postcss": "^8.4.41",
