+++ apps/website/src/app/links/discord/support/page.tsx
@@ -1,0 +1,5 @@
+import { redirect } from 'next/navigation';
+
+export default function Page() {
+	redirect('https://discord.gg/9FHagm6343');
+}
