+++ package.json
@@ -14,10 +14,10 @@
 		"prettier": "prettier --write \"apps/**/*.{ts,tsx,js,jsx}\" \"packages/**/*.{ts,tsx,js,jsx}\""
 	},
 	"devDependencies": {
-		"eslint": "^9.9.0",
+		"eslint": "^9.9.1",
 		"prettier": "^3.3.3",
 		"prettier-plugin-tailwindcss": "^0.6.6",
-		"turbo": "^2.0.14"
+		"turbo": "^2.1.1"
 	},
 	"packageManager": "pnpm@9.7.0"
 }
