+++ packages/database/package.json
@@ -23,6 +23,6 @@
 	},
 	"devDependencies": {
 		"@ticketer/eslint-config": "workspace:*",
-		"drizzle-kit": "^0.24.0"
+		"drizzle-kit": "^0.24.2"
 	}
 }
