+++ packages/djs-framework/package.json
@@ -17,7 +17,7 @@
 	},
 	"devDependencies": {
 		"@ticketer/eslint-config": "workspace:*",
-		"@types/node": "^22.4.1",
-		"tsx": "^4.17.0"
+		"@types/node": "^22.5.1",
+		"tsx": "^4.19.0"
 	}
 }
