+++ packages/eslint-config/package.json
@@ -6,15 +6,15 @@
 	"devDependencies": {
 		"@eslint/compat": "^1.1.1",
 		"@eslint/eslintrc": "^3.1.0",
-		"@eslint/js": "^9.9.0",
+		"@eslint/js": "^9.9.1",
 		"@types/eslint__js": "^8.42.3",
-		"eslint-config-next": "^14.2.5",
+		"eslint-config-next": "^14.2.7",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-config-turbo": "^2.0.14",
+		"eslint-config-turbo": "^2.1.1",
 		"eslint-plugin-drizzle": "^0.2.3",
 		"eslint-plugin-prettier": "^5.2.1",
 		"eslint-plugin-react": "^7.35.0",
 		"eslint-plugin-unicorn": "^55.0.0",
-		"typescript-eslint": "^8.2.0"
+		"typescript-eslint": "^8.3.0"
 	}
 }
